s32g: ddr: Add S32CTv1.4 Update 2 code
Added the DDR code from the latest Configuration Tool
version for S32G2/G3/R45 platforms.

Issue: ALB-8104

+++ drivers/nxp/s32g/ddr/ddr_init.c
@@ -204,7 +204,7 @@ static uint32_t load_phy_image(uint32_t start_addr, size_t size,
 
 	for (i = 0; i < size; i++) {
 		mmio_write_32((uintptr_t)current_addr, image[i]);
-		current_addr += sizeof(uint32_t);
+		current_addr += (uint32_t)sizeof(uint32_t);
 	}
 	return NO_ERR;
 }
